pyproject.toml
[tool.poetry]
name = "upk"
version = "0.1.0"
description = ""
authors = ["Cologler <skyoflw@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.9.3"
requests = "^2.25.0"
progressbar2 = "^3.53.1"
atomicwrites = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
